Main/SystemTableBuilder.cpp: size_t -> uint32_t warning. jiraref:VBP-1171
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@106598 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/SystemTableBuilder.cpp
@@ -455,8 +455,8 @@ int SystemTableBuilderAcpi::dumpTables(const char *pszFilename)
 
 int SystemTableBuilderAcpi::buildMadt(RTVFSIOSTREAM hVfsIos, size_t *pcbMadt)
 {
-    uint8_t abMadt[_4K];
-    size_t cbMadt = 0;
+    uint8_t  abMadt[_4K];
+    uint32_t cbMadt = 0;
 
     RT_ZERO(abMadt);
